requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pywebview==4.4.1

openai==1.3.0
openai-whisper

pyttsx3==2.90
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

pytesseract==0.3.10
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
python-dateutil==2.8.2
requests==2.31.0

numpy==1.26.4  # Compatible with librosa, soundfile, TensorFlow, contourpy
Pillow==10.0.1
